Work in progress change to how games are rendered
