+++ app/src/main/java/com/example/scorekeeper/AppUiState.kt
@@ -1,10 +1,10 @@
 package com.example.scorekeeper
 
+import com.example.scorekeeper.game.renderers.GameRenderer
 import com.example.scorekeeper.game.types.Game
 
 data class AppUiState(
     val gameCards: List<Game>,
     val isCreatingGame: Boolean,
-    val focusedGame: Game?,
-    val isFocusedGameVisible: Boolean
+    val gameRenderer: GameRenderer<Game>?
 )
