+++ app/src/main/java/com/example/scorekeeper/game/GameStateManager.kt
@@ -5,10 +5,9 @@ import androidx.datastore.preferences.core.edit
 import androidx.datastore.preferences.core.stringPreferencesKey
 import com.example.scorekeeper.dataStore
 import com.example.scorekeeper.game.types.Game
-import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 
-class GameManager(private val context: Context) {
+class GameStateManager(private val context: Context) {
     private val dataStore = context.dataStore;
     private val GAME_SAVES = stringPreferencesKey("game_saves")
